# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Lots Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/lots/id/{lot_id}": {
            "get": {
                "description": "Get Lot by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lots"
                ],
                "summary": "Get Lot by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Lot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseHTTP.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responseHTTP.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseHTTP.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Car": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "brandID": {
                    "type": "integer"
                },
                "carID": {
                    "type": "integer"
                },
                "engine": {
                    "type": "string"
                },
                "madeYear": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "modelID": {
                    "type": "integer"
                },
                "transmission": {
                    "type": "string"
                },
                "wheelDrive": {
                    "type": "string"
                }
            }
        },
        "domain.Lot": {
            "type": "object",
            "properties": {
                "car": {
                    "$ref": "#/definitions/domain.Car"
                },
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isLiked": {
                    "type": "boolean"
                },
                "lotID": {
                    "type": "integer"
                },
                "mileage": {
                    "type": "integer"
                },
                "postDate": {
                    "type": "string"
                },
                "salePrice": {
                    "type": "integer"
                },
                "saleStatus": {
                    "type": "string"
                },
                "sellerID": {
                    "type": "integer"
                },
                "vinCode": {
                    "type": "string"
                }
            }
        },
        "responseHTTP.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api"
    }
}